backend/api-service/pyproject.toml
[project]
name = "infusion-platform-api"
version = "0.1.0"
description = "Backend API service for closed-loop infusion platform"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Clinical Automation Team" }
]
dependencies = [
  "fastapi==0.110.0",
  "uvicorn[standard]==0.29.0",
  "sqlalchemy[asyncio]==2.0.28",
  "asyncpg==0.29.0",
  "alembic==1.13.1",
  "pydantic==2.6.3",
  "python-jose==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "pydantic-settings==2.2.1",
  "httpx==0.27.0"
]

[project.optional-dependencies]
test = [
  "pytest==8.1.1",
  "pytest-asyncio==0.23.5",
  "httpx==0.27.0",
  "respx==0.20.2"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
